+++ nv/mappings.py
@@ -16,12 +16,14 @@
 # along with NeoVintageous.  If not, see <https://www.gnu.org/licenses/>.
 
 import logging
+import re
 import traceback
 
 from NeoVintageous.nv import plugin
 from NeoVintageous.nv.settings import get_mode
 from NeoVintageous.nv.settings import get_partial_sequence
 from NeoVintageous.nv.settings import is_plugin_enabled
+from NeoVintageous.nv.utils import get_file_type
 from NeoVintageous.nv.variables import expand_keys
 from NeoVintageous.nv.vi import keys
 from NeoVintageous.nv.vi.cmd_base import ViMissingCommandDef
@@ -55,13 +57,37 @@ class Mapping:
         self.rhs = rhs
 
 
-def _find_partial_matches(mode: str, lhs: str) -> list:
-    return [x for x in _mappings[mode] if x.startswith(lhs)]
+def _find_partial_matches(view, mode: str, lhs: str) -> list:
+    matches = []
+    for map_lhs, map_rhs in _mappings[mode].items():
+        if isinstance(map_rhs, str):
+            if map_lhs.startswith(lhs):
+                matches.append(map_lhs)
+        else:
+            file_type = get_file_type(view)
+            if file_type and file_type in map_rhs:
+                if map_lhs.startswith(lhs):
+                    matches.append(map_lhs)
+            elif '' in map_rhs:
+                if map_lhs.startswith(lhs):
+                    matches.append(map_lhs)
+
+    return matches
 
 
-def _find_full_match(mode: str, lhs: str):
-    if lhs in _mappings[mode]:
-        return _mappings[mode][lhs]
+def _find_full_match(view, mode: str, lhs: str):
+    rhs = _mappings[mode].get(lhs)
+    if rhs:
+        if isinstance(rhs, str):
+            return rhs
+
+        try:
+            return _mappings[mode][lhs][get_file_type(view)]
+        except KeyError:
+            try:
+                return _mappings[mode][lhs]['']
+            except KeyError:
+                pass
 
 
 def _normalise_lhs(lhs: str) -> str:
@@ -73,6 +99,26 @@ def _normalise_lhs(lhs: str) -> str:
 
 
 def mappings_add(mode: str, lhs: str, rhs: str) -> None:
+    if re.match('^FileType$', lhs):
+        parsed = re.match('^([^ ]+) ([^ ]+)\\s+', rhs)
+        if parsed:
+            file_type = parsed.group(1)
+            file_type_lhs = parsed.group(2)
+            file_type_rhs = rhs[len(parsed.group(0)):]
+
+            file_type_lhs_norm = _normalise_lhs(file_type_lhs)
+
+            match = _mappings[mode].get(file_type_lhs_norm)
+
+            if not match:
+                _mappings[mode][file_type_lhs_norm] = {}
+            elif isinstance(match, str):
+                _mappings[mode][file_type_lhs_norm] = {'': match}
+
+            _mappings[mode][file_type_lhs_norm][file_type] = file_type_rhs
+
+            return
+
     _mappings[mode][_normalise_lhs(lhs)] = rhs
 
 
@@ -89,10 +135,10 @@ def mappings_is_incomplete(view) -> bool:
     mode = get_mode(view)
     seq = get_partial_sequence(view)
 
-    if _find_full_match(mode, seq):
+    if _find_full_match(view, mode, seq):
         return False
 
-    if _find_partial_matches(mode, seq):
+    if _find_partial_matches(view, mode, seq):
         return True
 
     return False
@@ -102,17 +148,18 @@ def mappings_can_resolve(view, key: str) -> bool:
     mode = get_mode(view)
     sequence = get_partial_sequence(view) + key
 
-    if _find_full_match(mode, sequence):
+    if _find_full_match(view, mode, sequence):
         return True
 
-    if _find_partial_matches(mode, sequence):
+    if _find_partial_matches(view, mode, sequence):
         return True
 
     return False
 
 
-def _seq_to_mapping(mode: str, seq: str):
-    full_match = _find_full_match(mode, seq)
+def _seq_to_mapping(view, seq: str):
+    mode = get_mode(view)
+    full_match = _find_full_match(view, mode, seq)
     if full_match:
         return Mapping(seq, full_match)
 
@@ -184,7 +231,7 @@ def mappings_resolve(view, sequence: str = None, mode: str = None, check_user_ma
         # let the definition handle any special-cases itself instead of passing
         # off the responsibility to the feed key command.
 
-        command = _seq_to_mapping(get_mode(view), seq)
+        command = _seq_to_mapping(view, seq)
 
     if not command:
         command = _seq_to_command(view, to_bare_command_name(seq), mode or get_mode(view))